refactor: 쿠폰 만들기 Mutation 로직 분리
