+++ frontend/src/@hooks/kkogkkog/useKkogKkogForm.ts
@@ -1,9 +1,5 @@
 import { ChangeEvent, FormEvent, useState } from 'react';
-import { useMutation } from 'react-query';
-import { useNavigate } from 'react-router-dom';
 
-import { createKkogkkog } from '@/apis/kkogkkog';
-import { PATH } from '@/Router';
 import {
   KKOGKKOG_COLORS,
   KKOGKKOG_ENG_TYPE,
@@ -14,28 +10,16 @@ import {
 } from '@/types/client/kkogkkog';
 import { UserResponse } from '@/types/remote/response';
 
-export const useKkogKkogForm = () => {
-  const navigate = useNavigate();
+import { useCreateKkogKkogMutation } from '../@queries/kkogkkog';
 
+export const useKkogKkogForm = () => {
   const [receiverList, setReceiverList] = useState<UserResponse[]>([]);
   const [couponType, setCouponType] = useState<KKOGKKOG_ENG_TYPE>(kkogkkogType[0].engType);
   const [modifier, setModifier] = useState<KKOGKKOG_MODIFIERS>(kkogkkogModifiers[0]);
   const [color, setColor] = useState<KKOGKKOG_COLORS>(kkogkkogColors[0]);
   const [message, setMessage] = useState('');
 
-  /** Add Typing */
-  const { mutate: createKkogKKogMutate } = useMutation(createKkogkkog, {
-    onSuccess() {
-      navigate(PATH.LANDING, {
-        state: {
-          action: 'create',
-        },
-      });
-    },
-    onError() {
-      alert('입력창을 확인하고 다시 시도해주세요.');
-    },
-  });
+  const createKkogKKogMutate = useCreateKkogKkogMutation();
 
   const onSelectType = (type: KKOGKKOG_ENG_TYPE) => {
     setCouponType(type);
@@ -78,7 +62,7 @@ export const useKkogKkogForm = () => {
       return;
     }
 
-    createKkogKKogMutate({
+    createKkogKKogMutate.mutate({
       receivers: receiverList.map(({ id }) => id),
       backgroundColor: color,
       modifier,
